+++ src/controllers/AuthController.js
@@ -15,11 +15,6 @@ Router.post("/emailcheck", AuthHandler.isNotLoggedIn, AuthServices.EmailCheck);
 Router.get("/", AuthServices.GetUser);
 Router.post("/emailauth", AuthHandler.isNotLoggedIn, AuthServices.EmailAuth);
 Router.post("/authcheck", AuthHandler.isNotLoggedIn, AuthServices.AuthCheck);
-Router.get("/google",AuthServices.GoogleLogin);
-Router.get("/google/callback",AuthServices.GoogleCallback,function(req,res){
-    res.send("콜백함수 체크");
-});
-Router.post("/socialnickname",AuthHandler.isLoggedIn,AuthServices.SocialNickname);
 
 
 export default Router;
